Beginning of big NERO refactor. environment is initializing properly
