+++ mods/NERO/environment.py
@@ -7,89 +7,6 @@ import constants
 import module
 import OpenNero
 
-
-class AgentState:
-    """
-    State that we keep for each agent
-    """
-    def __init__(self, agent):
-        self.id = agent.state.id
-        self.agent = agent
-        self.pose = (0, 0, 0)  # current x, y, heading
-        self.prev_pose = (0, 0, 0)
-        self.initial_position = OpenNero.Vector3f(0, 0, 0)
-        self.initial_rotation = OpenNero.Vector3f(0, 0, 0)
-        self.total_damage = 0
-        self.curr_damage = 0
-
-    def __str__(self):
-        x, y, h = self.pose
-        px, py, ph = self.prev_pose
-        return 'agent { id: %d, pose: (%.02f, %.02f, %.02f), prev_pose: (%.02f, %.02f, %.02f) }' % \
-            (self.id, x, y, h, px, py, ph)
-
-    def randomize(self):
-        dx = random.randrange(constants.SPAWN_RANGE * 2) - constants.SPAWN_RANGE
-        dy = random.randrange(constants.SPAWN_RANGE * 2) - constants.SPAWN_RANGE
-        self.initial_position.x = module.getMod().spawn_x[self.agent.get_team()] + dx
-        self.initial_position.y = module.getMod().spawn_y[self.agent.get_team()] + dy
-        self.prev_pose = self.pose = (self.initial_position.x,
-                                      self.initial_position.y,
-                                      self.initial_rotation.z)
-
-    def reset_pose(self, position, rotation):
-        self.prev_pose = self.pose = (position.x, position.y, rotation.z)
-
-    def update_damage(self):
-        """
-        Update the damage for an agent, returning the current damage.
-        """
-        self.total_damage += self.curr_damage
-        damage = self.curr_damage
-        self.curr_damage = 0
-        return damage
-
-    def update_pose(self, move_by, turn_by):
-        dist = constants.MAX_MOVEMENT_SPEED * move_by
-        heading = common.wrap_degrees(self.agent.state.rotation.z, turn_by)
-        x = self.agent.state.position.x + dist * math.cos(math.radians(heading))
-        y = self.agent.state.position.y + dist * math.sin(math.radians(heading))
-
-        self.prev_pose = self.pose
-        self.pose = (x, y, heading)
-
-        # try to update position
-        pos = copy.copy(self.agent.state.position)
-        pos.x = x
-        pos.y = y
-        self.agent.state.position = pos
-
-        # try to update rotation
-        rot = copy.copy(self.agent.state.rotation)
-        rot.z = heading
-        self.agent.state.rotation = rot
-
-class AgentTrace:
-    '''
-    Class for storing the trace of an agent
-    '''
-    def __init__(self):
-        self.initial_step = None
-        self.position = []
-        self.rotation = []
-        self.velocity = []
-        self.sensors = []
-        self.actions = []
-
-    def print_trace(self, filename):
-        with open(filename, 'w') as file:
-            for i in range(len(self.position)):
-                file.write(str([self.position[i].x, self.position[i].y, self.position[i].z]))
-                file.write(str([self.rotation[i].x, self.rotation[i].y, self.rotation[i].z]))
-            for i in range(len(self.sensors)):
-                file.write(str(self.sensors[i]))
-                file.write(str(self.actions[i]))
-
 class NeroEnvironment(OpenNero.Environment):
     """
     Environment for the Nero
@@ -100,53 +17,116 @@ class NeroEnvironment(OpenNero.Environment):
         """
         OpenNero.Environment.__init__(self)
 
+        self.flag_loc = None
+        self.flag_id = None
+
         self.lifetime = constants.DEFAULT_LIFETIME
         self.hitpoints = constants.DEFAULT_HITPOINTS
-        self.epsilon = constants.DEFAULT_EE / 100.0
 
-        self.curr_id = 0
-        self.max_steps = 20
-        self.MAX_DIST = math.hypot(constants.XDIM, constants.YDIM)
-
-        self.states = {}
-        self.teams = dict((t, set()) for t in constants.TEAMS)
-        self.agents_to_load = {}
-
+        x = constants.XDIM / 2.0
+        y = constants.YDIM / 3.0
+        self.spawn_x = {
+            constants.OBJECT_TYPE_TEAM_0: x,
+            constants.OBJECT_TYPE_TEAM_1: x
+        }
+        self.spawn_y = {
+            constants.OBJECT_TYPE_TEAM_0: y,
+            constants.OBJECT_TYPE_TEAM_1: 2 * y
+        }
         self.reward_weights = dict((f, 0.) for f in constants.FITNESS_DIMENSIONS)
-
-        # we only want to compute the center of mass of friends once per tick
-        self.friend_center = dict((team, None) for team in constants.TEAMS)
-        # we need to know when to update the friend center
-        self.friend_center_cache = dict((team, {}) for team in constants.TEAMS)
         
-        self.tracing = False             # indicates if the agent is being traced
-        self.simdisplay = True           # indicates if the simulation of agents is being displayed
-        self.trace = None                # trace of agents
-        self.use_trace = False           # use loaded trace for calculating reward
-        self.run_backprop = False        # whether to run backprop if trace is loaded
-        self.path_markers_trace = []     # ids of objects used to mark path of trace
-        self.path_markers_champ = []     # ids of objects used to mark path of champ
-        self.bbmarkers = []
+    def setup(self):
+        # world walls
+        height = constants.HEIGHT + constants.OFFSET
+        common.addObject(
+            "data/shapes/cube/Cube.xml",
+            OpenNero.Vector3f(constants.XDIM/2, 0, height),
+            OpenNero.Vector3f(0, 0, 90),
+            scale=OpenNero.Vector3f(constants.WIDTH, constants.XDIM, constants.HEIGHT*2),
+            label="World Wall0",
+            type=constants.OBJECT_TYPE_OBSTACLE)
+        common.addObject(
+            "data/shapes/cube/Cube.xml",
+            OpenNero.Vector3f(0, constants.YDIM/2, height),
+            OpenNero.Vector3f(0, 0, 0),
+            scale=OpenNero.Vector3f(constants.WIDTH, constants.YDIM, constants.HEIGHT*2),
+            label="World Wall1",
+            type=constants.OBJECT_TYPE_OBSTACLE)
+        common.addObject(
+            "data/shapes/cube/Cube.xml",
+            OpenNero.Vector3f(constants.XDIM, constants.YDIM/2, height),
+            OpenNero.Vector3f(0, 0, 0),
+            scale=OpenNero.Vector3f(constants.WIDTH, constants.YDIM, constants.HEIGHT*2),
+            label="World Wall2",
+            type=constants.OBJECT_TYPE_OBSTACLE)
+        common.addObject(
+            "data/shapes/cube/Cube.xml",
+            OpenNero.Vector3f(constants.XDIM/2, constants.YDIM, height),
+            OpenNero.Vector3f(0, 0, 90),
+            scale=OpenNero.Vector3f(constants.WIDTH, constants.XDIM, constants.HEIGHT*2),
+            label="World Wall3",
+            type=constants.OBJECT_TYPE_OBSTACLE)
+
+        # Add an obstacle wall in the middle
+        common.addObject(
+            "data/shapes/cube/Cube.xml",
+            OpenNero.Vector3f(constants.XDIM/2, constants.YDIM/2, height),
+            OpenNero.Vector3f(0, 0, 90),
+            scale=OpenNero.Vector3f(constants.WIDTH, constants.YDIM / 4, constants.HEIGHT*2),
+            label="World Wall4",
+            type=constants.OBJECT_TYPE_OBSTACLE)
+
+        # Add some trees
+        for i in (0.25, 0.75):
+            for j in (0.25, 0.75):
+                # don't collide with trees - they are over 500 triangles each
+                common.addObject(
+                    "data/shapes/tree/Tree.xml",
+                    OpenNero.Vector3f(i * constants.XDIM, j * constants.YDIM, constants.OFFSET),
+                    OpenNero.Vector3f(0, 0, 0),
+                    scale=OpenNero.Vector3f(1, 1, 1),
+                    label="Tree %d %d" % (10 * i, 10 * j),
+                    type=constants.OBJECT_TYPE_LEVEL_GEOM)
+                # collide with their trunks represented with cubes instead
+                common.addObject(
+                    "data/shapes/cube/Cube.xml",
+                    OpenNero.Vector3f(i * constants.XDIM, j * constants.YDIM, constants.OFFSET),
+                    OpenNero.Vector3f(0,0,0),
+                    scale=OpenNero.Vector3f(1,1,constants.HEIGHT),
+                    type=constants.OBJECT_TYPE_OBSTACLE)
+
+        # Add the surrounding Environment
+        common.addObject(
+            "data/terrain/NeroWorld.xml",
+            OpenNero.Vector3f(constants.XDIM/2, constants.YDIM/2, 0),
+            scale=OpenNero.Vector3f(1.2, 1.2, 1.2),
+            label="NeroWorld",
+            type=constants.OBJECT_TYPE_LEVEL_GEOM)
+
+    def remove_flag(self):
+        if self.flag_id:
+            common.removeObject(self.flag_id)
 
     def set_weight(self, key, value):
         self.reward_weights[key] = value
-        for team in self.teams:
-            rtneat = OpenNero.get_ai("rtneat-%s" % team)
-            if rtneat:
-                rtneat.set_weight(constants.FITNESS_INDEX[key], value)
 
-    def remove_all_agents(self, team):
-        for agent in list(self.teams[team]):
-            self.remove_agent(agent)
+    def change_flag(self, loc):
+        if self.flag_id:
+            common.removeObject(self.flag_id)
+        self.flag_loc = OpenNero.Vector3f(*loc)
+        self.flag_id = common.addObject(
+            "data/shapes/cube/BlueCube.xml",
+            self.flag_loc,
+            label="Flag",
+            scale=OpenNero.Vector3f(1, 1, 10),
+            type=constants.OBJECT_TYPE_FLAG)
 
-    def remove_agent(self, agent):
-        common.removeObject(agent.state.id)
-        try:
-            self.teams[agent.get_team()].discard(agent)
-            if agent in self.states:
-                self.states.pop(agent)
-        except:
-            pass
+    def place_basic_turret(self, loc):
+        return common.addObject(
+            "data/shapes/character/steve_basic_turret.xml",
+            OpenNero.Vector3f(*loc),
+            type=constants.OBJECT_TYPE_TEAM_1)
+
 
     def reset(self, agent):
         """
@@ -200,18 +180,7 @@ class NeroEnvironment(OpenNero.Environment):
         """
         if agent not in self.states:
             self.states[agent] = AgentState(agent)
-            self.teams[agent.get_team()].add(agent)
         return self.states[agent]
-
-    def getFriendFoe(self, agent):
-        """
-        Returns sets of all friend agents and all foe agents.
-        """
-        my_team = agent.get_team()
-        other_team = constants.OBJECT_TYPE_TEAM_1
-        if my_team == other_team:
-            other_team = constants.OBJECT_TYPE_TEAM_0
-        return self.teams[my_team], self.teams[other_team]
 
     def target(self, agent):
         """
@@ -270,36 +239,6 @@ class NeroEnvironment(OpenNero.Environment):
         """
         2A step for an agent
         """
-        # if this agent has a serialized representation waiting, load it.
-        chunk = self.agents_to_load.get(agent.state.id)
-        if chunk is not None:
-            print 'loading agent', agent.state.id, 'from', len(chunk), 'bytes'
-            del self.agents_to_load[agent.state.id]
-            try:
-                agent.from_string(chunk)
-            except:
-                # if loading fails, remove this agent.
-                print 'error loading agent', agent.state.id
-                self.remove_agent(agent)
-
-                # if a user has a badly formatted q-learning agent in a mixed
-                # population file, the agent won't load and will be properly
-                # removed here. however, RTNEAT has only allocated enough brainz
-                # to cover (pop_size - num_qlearning_agents) agents, so whenever
-                # it comes time to spawn new agents, RTNEAT will think that it
-                # needs to spawn an extra agent to cover for this "missing" one.
-                # to prevent this exception, we decrement pop_size here.
-                #
-                # this probably prevents teams from having the proper number of
-                # agents if the user clicks on the deploy button after loading a
-                # broken pop file ... but that's tricky to fix.
-                constants.pop_size -= 1
-
-                return agent.info.reward.get_instance()
-
-        # set the epsilon for this agent, in case it's changed recently.
-        agent.epsilon = self.epsilon
-
         state = self.get_state(agent)
 
         #Initilize Agent state
@@ -315,9 +254,6 @@ class NeroEnvironment(OpenNero.Environment):
         # display agent info if neccessary
         if hasattr(agent, 'set_display_hint'):
             agent.set_display_hint()
-
-        # spawn more agents if possible.
-        self.maybe_spawn(agent)
 
         # get the desired action of the agent
         move_by = action[constants.ACTION_INDEX_SPEED]
@@ -466,48 +402,10 @@ class NeroEnvironment(OpenNero.Environment):
 
         return reward
 
-    def maybe_spawn(self, agent):
-        '''Spawn more agents if there are more to spawn.'''
-        if agent.ai != 'rtneat' or agent.group != 'Agent':
-            return
-
-        team = agent.get_team()
-        rtneat = OpenNero.get_ai('rtneat-%s' % team)
-        if not rtneat or not rtneat.ready():
-            return
-
-        friends, foes = self.getFriendFoe(agent)
-        if len(friends) >= constants.pop_size:
-            return
-
-        if agent is tuple(f for f in friends if f.ai == agent.ai)[0]:
-            module.getMod().spawnAgent(team=team, ai=agent.ai)
-
     def sense(self, agent, observations):
         """
         figure out what the agent should sense
         """
-        # the last observation is whether there is a target for the agent
-        # update: this is now being done using a RaySensor
-        #observations[constants.SENSOR_INDEX_TARGETING[0]] = int(self.target(agent) is not None)
-
-#        friends, foes = self.getFriendFoe(agent)
-#        if friends:
-#            # the 2 before that are the angle and heading to the center of mass of
-#            # the agent's team
-#            ax, ay = agent.state.position.x, agent.state.position.y
-#            cx, cy = self.get_friend_center(agent, friends)
-#            fd = self.distance((ax, ay), (cx, cy))
-#            ah = agent.state.rotation.z
-#            fh = self.angle((ax, ay, ah), (cx, cy)) + 180.0
-#            if fd <= constants.MAX_FRIEND_DISTANCE:
-#                observations[constants.SENSOR_INDEX_FRIEND_RADAR[0]] = fd / 15.0
-#                observations[constants.SENSOR_INDEX_FRIEND_RADAR[1]] = fh / 360.0
-
-
-        # The code above miscalculates the friend center by looking at all teammates
-        # instead of looking only inside the friend radius.
-        # Updated friend sensing code:
         my_team = agent.get_team()
         all_friends = self.teams[my_team] 
 
@@ -530,43 +428,7 @@ class NeroEnvironment(OpenNero.Environment):
                 value = max(0, min(value, 1))
                 observations[constants.SENSOR_INDEX_FRIEND_RADAR[0]] = value
                 observations[constants.SENSOR_INDEX_FRIEND_RADAR[1]] = fh / 360.0
-
-        if self.tracing:
-            if self.trace.initial_step is None:
-                self.trace.initial_step = agent.step
-            self.trace.position.append(agent.state.position)
-            self.trace.rotation.append(agent.state.rotation)
-            self.trace.sensors.append([o for o in observations])
-        
-        from agent import KeyboardAgent
-        if self.tracing and isinstance(agent, KeyboardAgent):
-            id = common.addObject(
-                'data/shapes/cube/YellowCube.xml',
-                position = agent.state.position,
-                scale = OpenNero.Vector3f(0.25, 0.25, 0.25),
-                type = constants.OBJECT_TYPE_LEVEL_GEOM)
-            self.path_markers_trace.append(id)
-
         return observations
-
-    def get_friend_center(self, agent, friends):
-        ''' get the x, y position of the center of mass of the friends group '''
-        id = agent.state.id
-        step = agent.step
-        team = agent.get_team()
-        # only recompute the center of mass once we find an agent that goes
-        # to the next step since the last computation
-        if self.friend_center[team] is None or step != self.friend_center_cache[team].get(id, step):
-            n = float(len(friends))
-            cx, cy = 0, 0
-            for f in friends:
-                fx, fy = f.state.position.x, f.state.position.y
-                cx += fx / n
-                cy += fy / n
-            self.friend_center[team] = (cx, cy)
-            self.friend_center_cache[team].clear()
-        self.friend_center_cache[team][id] = step
-        return self.friend_center[team]
 
     def distance(self, a, b):
         """
@@ -619,7 +481,6 @@ class NeroEnvironment(OpenNero.Environment):
         if agent.group == 'Turret':
             return False
 
-        team = agent.get_team()
         state = self.get_state(agent)
         dead = self.hitpoints > 0 and state.total_damage >= self.hitpoints
         old = self.lifetime > 0 and agent.step > 0 and 0 == agent.step % self.lifetime
@@ -634,10 +495,7 @@ class NeroEnvironment(OpenNero.Environment):
                 agent.teleport()
             return False
 
-        rtneat = OpenNero.get_ai("rtneat-%s" % team)
-        orphaned = rtneat and not rtneat.has_organism(agent)
-
-        return orphaned or dead or old
+        return dead or old
     
     def get_hitpoints(self, agent):
         damage = self.get_state(agent).total_damage
@@ -652,62 +510,3 @@ class NeroEnvironment(OpenNero.Environment):
         """
         common.killScript((constants.MENU_JAR, constants.MENU_CLASS))
         return True
-    
-    def start_tracing(self):
-        '''
-        create trace object and start tracing
-        tracing of only one agent is currently supported
-        '''
-        self.trace = AgentTrace()
-        self.tracing = True
-
-    def stop_tracing(self):
-        '''
-        stop tracing the agent
-        '''
-        self.tracing = False
-
-    def save_trace(self, filename):
-        '''
-        pickle the current trace in the given file
-        '''
-        if not self.trace:
-            print 'empty trace on save'
-        with open(filename, 'w') as file:
-            pickle.dump(self.trace, file)
-
-    def use_demonstration(self):
-        '''
-        Start using the currently loaded demonstration
-        '''
-        if self.trace is not None:
-            print 'starting to use demonstration'
-            self.use_trace = True
-        else:
-            print 'no demonstration to use'
-            
-    def load_trace(self, filename):
-        '''
-        load previously pickled trace from the given file
-        '''
-        with open(filename, 'r') as file:
-            self.trace = pickle.load(file)
-        self.use_trace = True
-        for pos in self.trace.position:
-            id = common.addObject(
-                "data/shapes/cube/YellowCube.xml",
-                position = pos,
-                scale = OpenNero.Vector3f(0.25,0.25,0.25),
-                type = constants.OBJECT_TYPE_LEVEL_GEOM)
-            self.path_markers_trace.append(id)
-
-    def cancel_demonstration(self):
-        '''
-        unload previously loaded trace
-        '''
-        self.tracing = False
-        self.trace = None
-        self.use_trace = False
-        while len(self.path_markers_trace) > 0:
-            id = self.path_markers_trace.pop()
-            common.removeObject(id)
